springboot整合 rabbitmq

+++ src/main/java/com/springboot/girl/bean/Order.java
@@ -18,9 +18,9 @@ public class Order {
 
     private BigDecimal goodsPrice;
 
-    private Byte orderChannel;
+    private Integer orderChannel;
 
-    private Byte status;
+    private Integer status;
 
     private Date createDate;
 
@@ -82,19 +82,19 @@ public class Order {
         this.goodsPrice = goodsPrice;
     }
 
-    public Byte getOrderChannel() {
+    public Integer getOrderChannel() {
         return orderChannel;
     }
 
-    public void setOrderChannel(Byte orderChannel) {
+    public void setOrderChannel(Integer orderChannel) {
         this.orderChannel = orderChannel;
     }
 
-    public Byte getStatus() {
+    public Integer getStatus() {
         return status;
     }
 
-    public void setStatus(Byte status) {
+    public void setStatus(Integer status) {
         this.status = status;
     }
